refactored components so that errors can be shown appropriately
