+++ utils/aws-S3-Client.ts
@@ -1,0 +1,6 @@
+import { S3Client } from "@aws-sdk/client-s3";
+// Set the AWS Region.
+const REGION = process.env.REGION;
+// Create an Amazon S3 service client object.
+const s3Client = new S3Client({ region: REGION });
+export { s3Client };
